refactor package structure of value ast node
