+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/main/java/org/apache/shardingsphere/sql/parser/sql/value/identifier/IdentifierValue.java
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.value.impl;
+package org.apache.shardingsphere.sql.parser.sql.value.identifier;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.sql.parser.sql.value.ValueASTNode;
 
 /**
- * Literal value.
+ * Identifier value.
  *
  * @author panjuan
  */
 @RequiredArgsConstructor
 @Getter
-public final class LiteralValue implements ValueASTNode<String> {
+public final class IdentifierValue implements ValueASTNode<String> {
     
     private final String value;
 }
